backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-docx==1.1.2
requests==2.32.3
anthropic==0.34.2
python-dotenv==1.0.1
python-multipart==0.0.9
